pro-websocket/requirements.txt
altgraph==0.16.1
future==0.17.1
macholib==1.11
pefile==2018.8.8
PyInstaller==3.4
pyserial==3.4
pywin32==224
pywin32-ctypes==0.2.0
tornado==5.1.1
